refactor: extracted duplication between StepRunner and TeardownRunner into abstract XbehaveTestRunner
