+++ src/Xbehave.2.Execution/StepRunner.cs
@@ -6,18 +6,15 @@ namespace Xbehave.Execution
 {
     using System;
     using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
     using System.Linq;
     using System.Reflection;
-    using System.Security;
     using System.Threading;
     using System.Threading.Tasks;
-    using Xbehave.Execution.Shims;
     using Xbehave.Sdk;
     using Xunit.Abstractions;
     using Xunit.Sdk;
 
-    public class StepRunner : TestRunner<IXunitTestCase>
+    public class StepRunner : XbehaveTestRunner
     {
         private readonly string stepDisplayName;
         private readonly Step step;
@@ -62,70 +59,22 @@ namespace Xbehave.Execution
             get { return this.teardowns.ToArray(); }
         }
 
-        protected override async Task<Tuple<decimal, string>> InvokeTestAsync(ExceptionAggregator aggregator)
+        protected override async Task RunTestAsync()
         {
-            var output = string.Empty;
-            var testOutputHelper = ConstructorArguments.OfType<TestOutputHelper>().FirstOrDefault();
-            if (testOutputHelper != null)
-            {
-                testOutputHelper.Initialize(this.MessageBus, this.Test);
-            }
-
-            var timer = new ExecutionTimer();
-            var oldSyncContext = SynchronizationContext.Current;
-
             try
             {
-                var asyncSyncContext = new AsyncTestSyncContext(oldSyncContext);
-                SetSynchronizationContext(asyncSyncContext);
-
-                await aggregator.RunAsync(
-                    () => timer.AggregateAsync(
-                        async () =>
-                        {
-                            var result = this.step.Body();
-                            var task = result as Task;
-                            if (task != null)
-                            {
-                                await task;
-                            }
-                            else
-                            {
-                                var ex = await asyncSyncContext.WaitForCompletionAsync();
-                                if (ex != null)
-                                {
-                                    aggregator.Add(ex);
-                                }
-                            }
-                        }));
+                var result = this.step.Body();
+                var task = result as Task;
+                if (task != null)
+                {
+                    await task;
+                }
             }
             finally
             {
-                SetSynchronizationContext(oldSyncContext);
+                this.teardowns.AddRange(this.step.Disposables.Select(disposable => (Action)disposable.Dispose));
+                this.teardowns.AddRange(this.step.Teardowns);
             }
-
-            this.teardowns.AddRange(this.step.Disposables.Select(disposable => (Action)disposable.Dispose));
-            this.teardowns.AddRange(this.step.Teardowns);
-
-            var executionTime = timer.Total;
-
-            if (testOutputHelper != null)
-            {
-                output = testOutputHelper.Output;
-                testOutputHelper.Uninitialize();
-            }
-
-            return Tuple.Create(executionTime, output);
-        }
-
-        [SuppressMessage(
-            "Microsoft.Security",
-            "CA2136:TransparencyAnnotationsShouldNotConflictFxCopRule",
-            Justification = "From xunit.")]
-        [SecuritySafeCritical]
-        private static void SetSynchronizationContext(SynchronizationContext context)
-        {
-            SynchronizationContext.SetSynchronizationContext(context);
         }
     }
 }
